build both versions

+++ sparkler-core/deploy.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
+standalone=$1
+
 pip install databricks-cli
 
-rm -rf build/sparkler-app-0.3.1-SNAPSHOT
-#rm -rf build/plugins
+if [ "$standalone" = "true" ]; then
+  ~/.local/bin/databricks fs cp --recursive --overwrite build/ dbfs:/FileStore/sparkler-standalone/
+else
+  rm -rf build/sparkler-app-0.3.1-SNAPSHOT
 
-~/.local/bin/databricks fs cp --recursive --overwrite build/ dbfs:/FileStore/sparkler-submit/
+  ~/.local/bin/databricks fs cp --recursive --overwrite build/ dbfs:/FileStore/sparkler-submit/
+fi
